- Index type was wrong, should be U32 instead of U16. - vec type was also wrong, std430 needs vec4s for arrays, vec3 gets aligned to 16 bytes anyway.

+++ src/pipeline/stage.rs
@@ -133,7 +133,7 @@ impl Stage {
                         command_buffer,
                         mesh_buffer.indices.buffer,
                         mesh_buffer.indices.offset,
-                        vk::IndexType::UINT16,
+                        vk::IndexType::UINT32,
                     );
                     ctx.device.cmd_draw_indexed(
                         command_buffer,
